Add files via upload
These files represent the first draft of my project. The project involves interactivity between the user and the computer, where the user attempts to guess number stored within the PC. The user has full control over the parameters, they can chose the min and max number, as well as the number of guesses they want. There needs to be much more improvement and refinement on this project, involving UI rework, organization of code, and adding and updating more features.